review: move AuthenticatedPolarisPrincipalImpl to `polaris-service`
